Use an ap_varbuf to hold incoming messages
Previously we realloc'd the message buffer once for every incoming
frame. This led to poor performance for very fragmented messages. By
switching to an ap_varbuf, which roughly doubles its allocation every
time it needs to grow, we cut the number of allocations significantly.

Thanks to Paul Spangler for the suggestion for this fix!