handle_frame: parse from the start of the block
Have handle_incoming adjust the block pointer and block_size so we don't
have to pass a block_offset around between functions.

This exposes an oddity in the logic -- if an error is encountered before
any bytes are parsed, the framing state will be set to
DATA_FRAMING_CLOSE, but the block will not be advanced (because the
number of handled bytes is zero). Instead, the outer loop will call
handle_frame() *again*, at which point all the bytes will be marked as
"handled".